requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alpaca-trade-api==3.2.0
anyio==4.11.0
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.14.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==42.0.8
ctrader_open_api==0.9.2
curl_cffi==0.13.0
deprecation==2.1.0
frozendict==2.4.6
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inputimeout==1.0.4
msgpack==1.0.3
multidict==6.6.4
multitasking==0.0.12
numpy==2.3.3
packaging==25.0
pandas==2.3.3
peewee==3.18.2
platformdirs==4.4.0
propcache==0.3.2
protobuf==3.20.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-telegram-bot==22.5
pytz==2025.2
PyYAML==6.0.1
requests==2.32.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
Twisted==24.3.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
websocket-client==1.8.0
websockets==10.4
yarl==1.20.1
yfinance==0.2.58
zope.interface==8.1.1
openai>=1.0.0
